+++ src/main/java/core/basesyntax/Square.java
@@ -1,9 +1,9 @@
 package core.basesyntax;
 
 public class Square extends Figure {
-    private double side;
+    private int side;
 
-    public Square(Color color, double side) {
+    public Square(Color color, int side) {
         super(color);
         this.side = side;
     }
@@ -12,7 +12,7 @@ public class Square extends Figure {
         return side;
     }
 
-    public void setSide(double side) {
+    public void setSide(int side) {
         this.side = side;
     }
